Use std::experimental::string_view instead of std::string where possible

+++ Texture Atlas Generator/Utils/terminal color.hpp
@@ -9,14 +9,10 @@
 #ifndef engine_utils_terminal_color_hpp
 #define engine_utils_terminal_color_hpp
 
-#include <cstdio>
 #include <cstdint>
-#include <cassert>
-
-//ANSI escape codes!
 
 namespace Term {
-  enum class Color : uint8_t {
+  enum class Color : unsigned {
     BLACK,
     RED,
     GREEN,
